Make _abstract_from_html more generally applicable
For a better modular design, the function could take the chapter name as
a parameter and search the chapter text for it. This allows better
reusability for the function.

Additionally, check that the abstract text fetch does indeed return
something before trying to preprocess the text. No real point in
preprocessing Nones...

Addresses: Task 1