Add function for summarization
Go through the high frequency words and start going through the text
checking whether the word is included. First the title is checked and
then the sentences of the text. If a sentence isn't already in the
summary, it's added there. Otherwise it is skipped and the sentence
checking for the word is continued.

Include the title in the summary, if it contains a high frequency word.

Include the sentence in the summary if...
* A high frequency word is included and it's in the abstract.
* A high frequency word and a named entity is included.
* If none of the above apply, include the first sentence to include the
  high frequency word.

The raw (unprocessed) sentences are needed as they're the sentences
included into the summary if the stemmed versions of them are selected.
The chapter where the sentence appears in needs to be included for
applying the summarization logic from above. High frequency words are
required as those are the core of the summarization logic. Named
entities are also required for the summarization logic. Title is needed
both as raw and as stemmed for same reasons as the sentences.

The sentence ordering is stored using indexes so that the selected
sentences can be reordered to match the order of appearance within the
text.

Addresses: Task 2 & 3