Major datastructure revamp
Get rid of raw text via html. Switch to using sequences of (header, X)
tuples that contain X as either str (full text blob) or sequences of str
(sentences) or sequences of sequences of text (words separated into
sequences).

There's a main division into sentences_by_chapters and
processed_wordlists_by_chapters. The former considers the text
depicted as tuples (header, sequence of str [sentences]) and the latter
considers the text depicted as tuples (header, sequence of sequence of
str [sentences split into words]). The latter has also been preprocessed
whereas the former is raw.

We're making the assumption that there is only one title in an html
document: we're taking the first one. The title is depicted as a tuple
of raw str and a sequence of its processed words.

Addresses: Task 3