Summarization logic refactoring
The check for whether the sentence is already included in the summary is
moved to before the validity checks. This simplifies the if-forest
structure. Additionally, the check for whether the high freq word is
included in the sentence is reversed so that if it isn't, the sentence
is skipped. This also contributes to the simplicity and readability.

Updated comment messages.

Addresses: Task 3