Improve summary's logging & comments
While the current implementation considers only high frequency words as
the keywords, a more future proof solution would be to consider them
more generally as keywords. This way the code documentation is in line
with changes to the types of keywords used: The keywords could be
something other than high frequency words.