Add a named entity filter
Add an optional filter for removing named entities from the list of
keywords. This allows that more topic related (as opposed to topic
entity related) words may be selected for the summarization logic. If
named entities are associated with these topic related words, the same
sentences are selected as previously. However, the difference comes from
situations where a large portion of the keywords would be named
entities. In these cases a wider array of sentences are potential for
the summary as the keyword-named entity summarization logic becomes more
meaningful (as opposed to the same word constituting for being both the
keyword and the named entity in the sentence under inspection).