Add interface for sumy's summarizers
To better evaluate our summarizer against previously existing ones,
there's a need to be able to use external summarizers for the same
document. We've chosen sumy as the external library for this.

Sumy's python API documentation is basically non-existent so using it
would've required quite a bit of work reading through the code. This
didn't seem like a straight-forward task so we decided to go for a more
quick & dirty solution and use sumy's CLI via the standard subprocess
module. Works like a charm.

New dependency: sumy (PyPI)
Addresses: Task 5