Add option to run the summary evaluation
Add a command line argument for running the summary evaluation. The
current implementation assumes the number of documents is specified
(using --evaluate) when the dataset source_type is selected.

The ROGUE-N metric implementation was changed to be more useful. The
division into ROUGE recall and ROUGE precision is made to get actually
useful metrics out of the summaries.

The dataset entries are expected to be in a local file, and to contain
source URLs separated by newlines. The format for reference summaries is
expected to be so that there's an IDX field and afterwards there can be
matching Facet's that are used as the reference summaries. A facet is
always expected to be for the latest found IDX.

Additionally, there's the expectation that there are 3 reference summary
files: noise.txt, high_abstraction.txt and low_abstraction.txt. The
unique IDXs are split among these files.

It's worth noting that the interface is not the most intuitive as the
system was never properly designed and the task requirements were merely
met one by one. Refactoring and restructuring needed!

Addresses: Task 6