Refactor: Add config.py module
Add config.py module for storing some of the configuration parameters.
This is mainly motivated by the need to have properly configured
reference summary file paths.

Some little refactoring is also made as an unused parameter is removed
and indentation is update to match PEP-8 in the request hdr init.