sumy_interface: SUMMARIZERS into a dict
Change the SUMMARIZERS constant from a list of strings into a dictionary
of string:strings. This allows that we store summarizer names separately
for internal logic and for GUI display. The CLI can operate with the
same ones as the internal logic, but it would be preferred to have them
named differently in the GUI.