Add a class for checkbox column
Add a custom TKinter widget that takes its parent and a list of strings
as an input. It adds itself to its parent and creates left-anchored
checkboxes from the list of strings. This provides an easier interface
for accessing the selections of multiple checkboxes.

The internal data structure within the class is a dictionary so that the
checkbox texts can be mapped into their values. There's a need to create
a new dictionary upon read as the data structure stores the checkbox
values as tkinter.IntVars and it would be useful to return them as
integers.