CheckboxColumn rework
Rename the arguments and variables to be more descriptive. Change the
used checkbox name-value mapping to use tuple names so that a different
name can be shown in GUI and returned when the box is selected. This
allows that the internal naming can be returned even when there's
different naming shown in the GUI. Also, only return selected checkboxes
rather than a dict of selection statuses.