Restructuring: Separate text processing from summarization
Keep a function that capsulates the entire process by only taking in a
config dict as a parameter and handling both the text extraction,
preprocessing, keyword extraction, summarization etc. However, spread
the text processing itself into separate functions.

This allows that the separate steps can be called from elsewhere and the
results of each step are visible to the caller instead of just the
resulting summary. This is necessary for the GUI to work as envisioned.