GUI: Integrate autosummary to autosummary_gui
Add the functionality for the summarize-button. For each of our "own"
summarizers, run the text processing, keyword extraction, named entity
tagging and finally the summarization logic.

Hox! Only the URL (and possibly the local file) are currently supported!
Dataset would require some extra processing, which is not currently
done. Additionally, the summary's evaluation functions in autosummary.py
should probably be split into chunks the same way as was done for the
normal summarization.