GUI: Add result box printing for summarization results
Print the summarization results for both "own" and sumy's summarizers.
For each summarizer it's relevant to print the name of the summarizer in
question and the extracted summary. If the source_type is dataset,
document id, reference summary and evaluation metrics can also be
printed. For "own" we also have access to the keywords and named
entities, which we can print.